+++ ninaAPI/WebService/V2/Equipment/Dome.cs
@@ -9,7 +9,6 @@
 
 #endregion "copyright"
 
-using ASCOM.Com.DriverAccess;
 using EmbedIO;
 using EmbedIO.Routing;
 using EmbedIO.WebApi;