+++ ninaAPI/WebService/V2/Equipment/Focuser.cs
@@ -1,4 +1,4 @@
-﻿#region "copyright"
+#region "copyright"
 
 /*
     Copyright © 2025 Christian Palm (christian@palm-family.de)
@@ -9,14 +9,19 @@
 
 #endregion "copyright"
 
+using Accord;
 using EmbedIO;
 using EmbedIO.Routing;
 using EmbedIO.WebApi;
 using Newtonsoft.Json;
+using NINA.Core.Locale;
 using NINA.Core.Utility;
 using NINA.Equipment.Equipment.MyFocuser;
+using NINA.Equipment.Interfaces;
 using NINA.Equipment.Interfaces.Mediator;
+using NINA.WPF.Base.Mediator;
 using NINA.WPF.Base.Utility.AutoFocus;
+using NINA.WPF.Base.ViewModel.Equipment.Focuser;
 using ninaAPI.Utility;
 using OxyPlot;
 using System;
@@ -117,8 +122,11 @@ namespace ninaAPI.WebService.V2
                 {
                     response = CoreUtility.CreateErrorTable(new Error("Focuser not connected", 409));
                 }
-                focuser.MoveFocuser(position, new CancellationTokenSource().Token);
-                response.Response = "Move started";
+                else
+                {
+                    focuser.MoveFocuser(position, new CancellationTokenSource().Token);
+                    response.Response = "Move started";
+                }
             }
             catch (Exception ex)
             {
@@ -152,11 +160,20 @@ namespace ninaAPI.WebService.V2
                     AutoFocusToken?.Cancel();
                     AutoFocusToken = new CancellationTokenSource();
 
-                    AdvancedAPI.Controls.AutoFocusFactory.Create().StartAutoFocus(
+                    var service = AdvancedAPI.Controls.WindowFactory.Create();
+
+                    var autofocus = AdvancedAPI.Controls.AutoFocusFactory.Create();
+                    service.Show(autofocus, Loc.Instance["LblAutoFocus"], System.Windows.ResizeMode.CanResize, System.Windows.WindowStyle.ToolWindow);
+
+                    autofocus.StartAutoFocus(
                         AdvancedAPI.Controls.FilterWheel.GetInfo().SelectedFilter,
                         AutoFocusToken.Token,
                         AdvancedAPI.Controls.StatusMediator.GetStatus()
-                    );
+                    ).ContinueWith(result =>
+                    {
+                        AdvancedAPI.Controls.ImageHistory.AppendAutoFocusPoint(result.Result);
+                        service.DelayedClose(TimeSpan.FromSeconds(10));
+                    });
                     response.Response = "Autofocus started";
                 }
             }
@@ -194,6 +211,32 @@ namespace ninaAPI.WebService.V2
                 else
                 {
                     response = CoreUtility.CreateErrorTable(new Error("No AF available", 500));
+                }
+            }
+            catch (Exception ex)
+            {
+                Logger.Error(ex);
+                response = CoreUtility.CreateErrorTable(CommonErrors.UNKNOWN_ERROR);
+            }
+
+            HttpContext.WriteToResponse(response);
+        }
+
+        [Route(HttpVerbs.Get, "/equipment/focuser/pins/reverse")]
+        public async Task FocuserLastAF([QueryField] bool reversing)
+        {
+            HttpResponse response = new HttpResponse();
+
+            try
+            {
+                if (AdvancedAPI.Controls.Focuser.GetType().HasMethod("SetReverse"))
+                {
+                    AdvancedAPI.Controls.Focuser.GetType().GetMethod("SetReverse").Invoke(AdvancedAPI.Controls.Focuser, [reversing]);
+                    response.Response = "Reverse set";
+                }
+                else
+                {
+                    response = CoreUtility.CreateErrorTable(new Error("Platform not supported", 500));
                 }
             }
             catch (Exception ex)