+++ ninaAPI/WebService/V2/WebSocketV2.cs
@@ -1,4 +1,4 @@
-﻿#region "copyright"
+#region "copyright"
 
 /*
     Copyright © 2025 Christian Palm (christian@palm-family.de)
@@ -109,7 +109,9 @@ namespace ninaAPI.WebService.V2
 
             response.Response = responseData;
 
-            Hashtable eventTable = responseData.DeepClone();
+            // Deep clone using JSON serialization (BinaryFormatter is disabled in .NET 8)
+            string json = JsonConvert.SerializeObject(responseData);
+            Hashtable eventTable = JsonConvert.DeserializeObject<Hashtable>(json);
             eventTable.Add("Time", time);
             HttpResponse Event = new HttpResponse() { Type = HttpResponse.TypeSocket, Response = eventTable };
             Events.Add(Event);