+++ Changelog.md
@@ -2,6 +2,30 @@
 
 The changes for each individual (beta) release can be seen in each [release](https://github.com/christian-photo/ninaAPI/releases). This changelog will only cover the fully released versions.
 
+## 2.2.14.0
+
+- Added `rotator/reverse` to reverse its direction
+- Added `rotator/set-mechanical-range` to set the rotators mechanical range (full, 180°, 90°)
+- Added `/astro-util/moon-separation` to calculate the moon seperation for given coordinates
+- Added `imageType` to the capture endpoint
+
+## 2.2.13.0
+
+- The `IMAGE-SAVED` event and `image-history` now contain the Filenames of the images
+- `image/{index}/solve` and `prepared-image/solve` added
+- `equipment/filterwheel/add-filter` and `equipment/filterwheel/remove-filter` added
+
+## 2.2.12.1
+
+- When no prepared image is available, the endpoint now returns a 404 error
+- The autofocus endpont `equipment/focuser/auto-focus` now correctly opens the autofocus window
+- The endpoint `profile/change-value` now also works with indexed properties
+- Fixed an issue where the following endpoints would execute the command even though the validation had failed:
+  - `/equipment/focuser/move`
+  - `/equipment/switch/set`
+  - `/equipment/rotator/move`
+  - `/equipment/rotator/move-mechanical`
+
 ## 2.2.12.0
 
 - Minimum application version was increased to NINA 3.2!